backend/requirements.txt
fastapi
uvicorn
pycryptodome
pydantic